requirements.txt
pytorch-lightning==0.9.0
test-tube==0.7.1
numpy==1.17.3
pretrainedmodels==0.7.4
fire==0.2.1
scipy==1.3.1
tensorboardX==1.9
scikit-learn>=0.23
pytorch-ignite
imgaug
pycountry_convert
geopy
seaborn
